Move code into components (#21)
