+++ frontend/package.json
@@ -15,6 +15,7 @@
     "@types/socket.io-client": "^1.4.34",
     "@typescript-eslint/eslint-plugin": "^4.14.2",
     "@typescript-eslint/parser": "^4.14.2",
+    "axios": "^0.21.1",
     "connected-react-router": "^6.8.0",
     "d3": "^6.2.0",
     "eslint": "^7.19.0",
