+++ frontend/public/settings.json
@@ -1,0 +1,7 @@
+{
+    "plugins":[
+        {
+            "name": "plugin1"
+        }
+    ]
+}
