+++ frontend/src/components/example.component.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import { StateType } from './state/state.types';
+import { StateType } from '../state/state.types';
 
 interface ExampleComponentProps {
     notifications: string[];
