+++ frontend/src/pages/mainAppBar.component.tsx
@@ -1,0 +1,34 @@
+import { AppBar, CssBaseline, Toolbar, Typography } from '@material-ui/core';
+import { createStyles, makeStyles } from '@material-ui/core/styles';
+import React from 'react';
+
+const useStyles = makeStyles(() =>
+    createStyles({
+        root: {
+            display: 'flex',
+        },
+        // appBar: {
+        //     width: `calc(100% - ${drawerWidth}px)`,
+        //     marginLeft: drawerWidth,
+        // },
+    }),
+);
+
+const MainAppBar = (): React.ReactElement => {
+    const classes = useStyles();
+
+    return (
+        <div className={classes.root}>
+            <CssBaseline />
+            <AppBar position="fixed">
+                <Toolbar>
+                    <Typography variant="h6" noWrap>
+                        Visualising Optimisation Data
+                    </Typography>
+                </Toolbar>
+            </AppBar>
+        </div>
+    );
+};
+
+export default MainAppBar;
